Change routing.key config option to routing.key.suffix based on Andrew's suggestion.
Without specifying anything, the default behavior is to use only the
table name for replicated tables and tablename.partitionColumnValue
for partitioned table.

User can specify any column in the table as the routing key suffix,
the value in that column will be joined with the table name by a dot
to form the routiing key.